+++ src/torchoutil/__init__.py
@@ -14,9 +14,3 @@ __version__ = "0.3.2"
 
 
 from .nn.functional import *
-
-# Redundant imports to avoid to import PyTorch modules
-from . import hub as hub  # isort:skip
-from . import nn as nn  # isort:skip
-from . import optim as optim  # isort:skip
-from . import utils as utils  # isort:skip