+++ src/torchoutil/nn/__init__.py
@@ -2,7 +2,3 @@
 # -*- coding: utf-8 -*-
 
 from .modules import *
-
-# Redundant imports to avoid to import PyTorch modules
-from . import functional as functional  # isort:skip
-from . import modules as modules  # isort:skip